+++ src/cone/tile/_api.py
@@ -174,13 +174,55 @@ class TileRenderer(object):
 
 @implementer(ITile)
 class Tile(object):
+    """Tile class.
+    """
 
-    def __init__(self, path, attribute, name):
-        self.name = name
-        self.path = path
-        self.attribute = attribute
+    name = None
+    """Name of this tile as string.
+    """
+
+    path = None
+    """Path to template related to this tile as string. Expected in the form
+    ``packagename:template.pt``.
+    """
+
+    attribute = 'render'
+    """Name of function used to render this tile if no path set. Defaults to
+    ``render``
+    """
+
+    def __init__(self, path=None, attribute=None, name=None):
+        """Construct tile.
+
+        @param path: Path to template as string.
+        @param attribute: Name of function used to render this tile if no path
+        set. Defaults to ``render``.
+        @param name: Name of this tile as string.
+        """
+        if self.name is not None:
+            self.name = name
+        if path is not None:
+            self.path = path
+        if attribute is not None:
+            self.attribute = attribute
 
     def __call__(self, model, request):
+        """Render tile.
+
+        * Calls ``prepare`` function
+        * Check ``show`` flag and returns empty string if ``True``
+        * Check template ``path`` and renders template with tile as context
+          if set
+        * Looks up function on self by ``attribute`` name and use it for tile
+          rendering if no template path set.
+        * Check whether redirection has been triggered some when while
+          processing and return empty string if so, otherwide return rendered
+          result.
+
+        @param model: tile related model
+        @param request: pyramid request
+        @return string: rendered result
+        """
         self.model = model
         self.request = request
         self.prepare()
@@ -198,12 +240,19 @@ class Tile(object):
 
     @property
     def show(self):
+        """Flag whether this tile should be displayed. Defaults to ``True``
+        """
         return True
 
     def prepare(self):
+        """Function which can be used for preparation if desired before
+        rendering this tile. Does nothing by default.
+        """
         pass
 
     def render(self):
+        """Default rendering function for this tile if no template path set.
+        """
         raise NotImplementedError('Base Tile does not implement ``render``')
 
     def redirect(self, redirect):
@@ -258,28 +307,29 @@ def _secure_tile(tile, permission, authn_policy, authz_policy, strict):
 
 
 # Registration
-def registerTile(name, path=None, attribute='render', interface=Interface,
+def registerTile(name=None, path=None, attribute=None, interface=Interface,
                  class_=Tile, permission='view', strict=True, _level=2):
-    """registers a tile.
+    """Registers a tile.
 
     ``name``
-        identifier of the tile (for later lookup).
+        Identifier of the tile (for later lookup). If name is None, it is
+        taken from given tile class_. If name still None, an exception
+        is raised.
 
     ``path``
-        either relative path to the template or absolute path or path prefixed
+        Either relative path to the template or absolute path or path prefixed
         by the absolute package name delimeted by ':'. If ``path`` is used
         ``attribute`` is ignored.
 
     ``attribute``
-        attribute on the given _class to be used to render the tile. Defaults to
-        ``render``.
+        Function name on the given class_ to be used to render the tile.
 
     ``interface``
         Interface or Class of the pyramid model the tile is registered for.
 
     ``class_``
-        Class to be used to render the tile. usally ``cone.tile.Tile`` or a
-        subclass of. Promises to implement ``cone.tile.ITile``.
+        Class which implements ``cone.tile.ITile``. Normally ``cone.tile.Tile``
+        or subclass of it.
 
     ``permission``
         Enables security checking for this tile. Defaults to ``view``. If set to
@@ -294,10 +344,19 @@ def registerTile(name, path=None, attribute='render', interface=Interface,
         is a bit special to make doctests pass the magic path-detection.
         you must never touch it in application code.
     """
+    if name is None:
+        name = class_.name
+
+    if name is None:
+        raise ValueError((
+            'Tile ``name`` must be either given at registration time '
+            'or set on given tile class: {}'
+        ).format(str(class_)))
+
     if path and not (':' in path or os.path.isabs(path)):
         path = '%s:%s' % (caller_package(_level).__name__, path)
 
-    tile = class_(path, attribute, name)
+    tile = class_(path=path, attribute=attribute, name=name)
 
     registry = get_current_registry()
     registered = registry.adapters.registered
@@ -342,10 +401,10 @@ class tile(object):
     """Decorator to register classes and functions as tiles.
     """
 
-    def __init__(self, name, path=None, attribute='render',
+    def __init__(self, name=None, path=None, attribute=None,
                  interface=Interface, permission='view',
                  strict=True, _level=2):
-        """ see ``registerTile`` for details on the other parameters.
+        """See ``registerTile`` for details on the other parameters.
         """
         self.name = name
         if path and not (':' in path or os.path.isabs(path)):
@@ -357,7 +416,7 @@ class tile(object):
         self.strict = strict
 
     def __call__(self, ob):
-        registerTile(self.name,
+        registerTile(name=self.name,
                      path=self.path,
                      attribute=self.attribute,
                      interface=self.interface,